+++ lib/widgets/settings/theme_selection.dart
@@ -35,7 +35,7 @@ class SettingsThemeSelection extends StatelessWidget {
         itemBuilder: (_, index) {
           final String name = app.themes.handledThemes[index];
           final ThemeData data = app.themes.all[name]!;
-          return ThemePreview(name: name.capitalizeFirst!.tr, data: data);
+          return ThemePreview(name: name, data: data);
         },
       ),
     );